+++ apps/docs/src/app/platform/api-files.ts
@@ -9,6 +9,13 @@ export const API_FILES = {
     button: ['ButtonComponent'],
     checkbox: ['CheckboxComponent'],
     checkboxGroup: ['CheckboxGroupComponent'],
+    dynamicPage: [
+        'DynamicPageComponent',
+        'DynamicPageTitleComponent',
+        'DynamicPageKeyInfoComponent',
+        'DynamicPageHeaderComponent',
+        'DynamicPageContentComponent'
+    ],
     infoLabel: ['InfoLabelComponent'],
     input: ['InputComponent'],
     link: ['LinkComponent'],
@@ -20,6 +27,17 @@ export const API_FILES = {
     searchField: ['SearchFieldComponent'],
     select: ['SelectPlatformComponent'],
     splitMenuButton: ['SplitMenuButtonComponent'],
+    table: [
+        'TableComponent',
+        'TableToolbarComponent',
+        'TableToolbarActionsComponent',
+        'TableColumnComponent',
+        'FdpTableHeader',
+        'FdpHeaderCellDef',
+        'FdpTableCell',
+        'FdpCellDef'
+
+    ],
     textarea: ['TextAreaComponent'],
     panel: [
         'PlatformPanelComponent',
@@ -34,5 +52,8 @@ export const API_FILES = {
     list: ['ListComponent', 'StandardListItemComponent', 'ListFooter', 'ListGroupHeader', 'ListConfig'],
     standardlistitem: ['ListComponent', 'StandardListItemComponent', 'ListFooter', 'ListGroupHeader', 'ListConfig'],
     thumbnail: ['ThumbnailComponent'],
-    objectAttribute: ['ObjectAttributeComponent']
+    objectAttribute: ['ObjectAttributeComponent'],
+    actionlistitem: ['ActionListItemComponent', 'ListComponent', 'ListConfig'],
+    displaylistitem: ['DisplayListItemComponent', 'ListComponent', 'ListFooter', 'ListGroupHeader', 'ListHeader', 'ListConfig'],
+    objectlistitem: ['ObjectListItemComponent', 'ListComponent', 'ListFooter', 'ListGroupHeader', 'ListHeader', 'ListConfig']
 };