+++ apps/docs/src/app/platform/component-docs/platform-thumbnail/platform-thumbnail-examples/platform-thumbnail-horizontal-example.component.ts
@@ -8,38 +8,38 @@ import { Media } from '@fundamental-ngx/platform';
 export class PlatformThumbnailHorizontalExampleComponent {
 
     data: Media[] = [{
-        thumbnailUrl: 'http://lorempixel.com/640/480/nature/2',
+        thumbnailUrl: 'http://picsum.photos/id/316/600',
         mediaType: 'image',
-        mediaUrl: 'http://lorempixel.com/640/480/nature/2',
-        alt: 'Failed to load http://lorempixel.com/640/480/nature/2',
+        mediaUrl: 'http://picsum.photos/id/316/600',
+        alt: 'Failed to load http://picsum.photos/id/316/600',
         label: 'nature'
     },
     {
-        thumbnailUrl: 'http://lorempixel.com/480/640/animals/2',
+        thumbnailUrl: 'http://picsum.photos/id/317/600',
         mediaType: 'image',
-        mediaUrl: 'http://lorempixel.com/480/640/animals/2',
-        alt: 'Failed to load http://lorempixel.com/480/640/animals/2',
+        mediaUrl: 'http://picsum.photos/id/317/600',
+        alt: 'Failed to load http://picsum.photos/id/317/600',
         label: 'animal'
     },
     {
-        thumbnailUrl: 'http://lorempixel.com/600/600/city/2',
+        thumbnailUrl: 'http://picsum.photos/id/318/600',
         mediaType: 'image',
-        mediaUrl: 'http://lorempixel.com/600/600/city/2',
-        alt: 'Failed to load http://lorempixel.com/600/600/city/2',
+        mediaUrl: 'http://picsum.photos/id/318/600',
+        alt: 'Failed to load http://picsum.photos/id/318/600',
         label: 'culture'
     },
     {
-        thumbnailUrl: 'http://lorempixel.com/640/480/sports/2',
+        thumbnailUrl: 'http://picsum.photos/id/319/600',
         mediaType: 'image',
-        mediaUrl: 'http://lorempixel.com/640/480/sports/2',
-        alt: 'Failed to load http://lorempixel.com/640/480/sports/2',
+        mediaUrl: 'http://picsum.photos/id/319/600',
+        alt: 'Failed to load http://picsum.photos/id/319/600',
         label: 'sports'
     },
     {
-        thumbnailUrl: 'http://lorempixel.com/400/400/abstract/2',
+        thumbnailUrl: 'http://picsum.photos/id/321/600',
         mediaType: 'image',
-        mediaUrl: 'http://lorempixel.com/400/400/abstract/2',
-        alt: 'Failed to load http://lorempixel.com/400/400/abstract/2',
+        mediaUrl: 'http://picsum.photos/id/321/600',
+        alt: 'Failed to load http://picsum.photos/id/321/600',
         label: 'art'
     }
     ];