+++ libs/platform/src/lib/components/table/enums/index.ts
@@ -1,0 +1,3 @@
+export * from './content-density.type';
+export * from './selection-mode.type';
+export * from './column-align.type';